Project Added:
The Level blockout of the small portion of a villa has been successfully added in the repo. This has been done for the Assignment 1's purpose of Technical Art (Term 4). The exterior part has been done with the usage of the created MODULAR ASSETS from MAYA software and imported it here.