Another Addition- Reworked MAYA files and Texture Imports:
Removed the old props and replaced them with reworked MAYA FBX files. Chair has to be reworked still for the scaling issue and that will be done in the next commit, other than that remaining looks fine. The texture part also will be done and pushed in the upcoming commit.